requirements.txt
runpod==1.7.7
torch==2.1.2
torchvision==0.16.2
transformers==4.41.2
diffusers==0.29.2
accelerate==0.30.1
safetensors==0.4.3
sentencepiece==0.2.0
Pillow==10.3.0
numpy==1.26.4
protobuf==4.25.3
huggingface_hub==0.23.4
